speech-service/requirements.txt
Flask
python-dotenv
assemblyai
flask-socketio
flask-cors
psycopg2-binary
flask-sqlalchemy 
openai
vosk
ffmpeg
numpy